import refactor

+++ Engine/Debuging/DebugShape.h
@@ -1,16 +1,15 @@
 ﻿#pragma once
-#include <SDL_stdinc.h>
 
 struct Color
 {
-    Color(const Uint8 r, const Uint8 g, const Uint8 b, const Uint8 a): r(r), g(g), b(b), a(a)
+    Color(const int r, const int g, const int b, const int a): r(r), g(g), b(b), a(a)
     {
     }
 
-    Uint8 r = 255;
-    Uint8 g = 255;
-    Uint8 b = 255;
-    Uint8 a = 255;
+    int r = 255;
+    int g = 255;
+    int b = 255;
+    int a = 255;
 };
 
 class DebugShape
